+++ src/decoder.rs
@@ -351,6 +351,7 @@ fn clone_frame_state(fs: &FrameState) -> FrameState {
         cdef_idx: fs.cdef_idx.clone(),
         cdef_sb_cols: fs.cdef_sb_cols,
         cdef_sb_rows: fs.cdef_sb_rows,
+        tx_types: fs.tx_types.clone(),
     }
 }